@@ -4,40 +4,25 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-brands-svg-icons": "^5.14.0",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.11",
-    "@types/bootstrap": "^5.0.1",
-    "@types/jest": "^26.0.13",
-    "@types/lodash": "^4.14.161",
-    "@types/node": "^14.10.1",
-    "@types/react": "^17.0.0",
-    "@types/react-helmet": "^6.1.0",
-    "@types/react-intl-redux": "^0.1.14",
-    "@types/reactstrap": "^8.5.1",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
     "bootstrap": "^4.5.2",
-    "commander": "^7.2.0",
-    "gatsby": "^3.3.0",
-    "gatsby-link": "^3.3.0",
-    "gatsby-plugin-react-helmet": "^4.3.0",
-    "gatsby-plugin-typescript": "^3.3.0",
-    "jest": "^26.4.2",
-    "lodash": "^4.17.20",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "gatsby": "^4.11.2",
+    "gatsby-link": "^4.11.1",
+    "gatsby-plugin-react-helmet": "^5.11.0",
+    "gatsby-plugin-typescript": "^4.11.1",
+    "i18next": "^21.6.14",
+    "i18next-browser-languagedetector": "^6.1.4",
+    "lodash": "^4.17.21",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-helmet": "^6.1.0",
-    "react-intl": "5.15.8",
-    "react-intl-redux": "^2.2.0",
-    "react-redux": "^7.2.1",
-    "react-twitter-widgets": "^1.9.5",
-    "reactstrap": "^8.5.1",
-    "redux": "^4.0.5",
-    "ts-jest": "^26.3.0",
-    "ts-node": "^9.0.0",
-    "tslint": "^6.1.3",
-    "typescript": "^4.0.2",
-    "url-parse": "^1.4.3"
+    "react-i18next": "^11.16.2",
+    "react-twitter-widgets": "^1.10.0",
+    "reactstrap": "^8.7.1",
+    "use-local-storage": "^2.3.6"
   },
   "keywords": [
     "gatsby"
@@ -52,7 +37,21 @@
     "lint": "tslint -p tsconfig.json"
   },
   "devDependencies": {
-    "prettier": "^2.1.1"
+    "@types/bootstrap": "^4.5.2",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.181",
+    "@types/node": "^17.0.23",
+    "@types/react": "^17.0.43",
+    "@types/react-helmet": "^6.1.5",
+    "@types/reactstrap": "^8.7.1",
+    "commander": "^9.1.0",
+    "jest": "^27.5.1",
+    "prettier": "^2.6.2",
+    "prettier-plugin-organize-imports": "^2.3.4",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.6.3"
   },
   "repository": {
     "type": "git",
