@@ -1,11 +1,11 @@
 import * as React from 'react'
-import { FormattedMessage, useIntl } from 'react-intl'
+import { useTranslation } from 'react-i18next'
 import { Table } from 'reactstrap'
 import AppLink from '../components/AppLink'
 import apps, { AppStatuses, IAppData } from '../data/apps'
 
 const appToTableRow: React.FC<IAppData> = (app) => {
-  const intl = useIntl()
+  const { t } = useTranslation()
   return (
     <tr key={app.title}>
       <th style={{ padding: '10px' }} className="align-middle" scope="row">
@@ -15,7 +15,7 @@ const appToTableRow: React.FC<IAppData> = (app) => {
       <td className="align-middle">
         <div
           dangerouslySetInnerHTML={{
-            __html: intl.formatMessage({ id: `app.${app.title}` }),
+            __html: t(`app.${app.title}`),
           }}
         />
       </td>
@@ -23,25 +23,24 @@ const appToTableRow: React.FC<IAppData> = (app) => {
   )
 }
 
-const ActiveApps = () => (
-  <Table size="sm">
-    <thead>
-      <tr>
-        <th></th>
-        <th>
-          <FormattedMessage id="app.title" />
-        </th>
-        <th>
-          <FormattedMessage id="app.description" />
-        </th>
-      </tr>
-    </thead>
-    <tbody>
-      {apps
-        .filter((app) => app.status === AppStatuses.active)
-        .map((app) => appToTableRow(app))}
-    </tbody>
-  </Table>
-)
+const ActiveApps = () => {
+  const { t } = useTranslation()
+  return (
+    <Table size="sm">
+      <thead>
+        <tr>
+          <th></th>
+          <th>{t(`app.title`)}</th>
+          <th>{t(`app.description`)}</th>
+        </tr>
+      </thead>
+      <tbody>
+        {apps
+          .filter((app) => app.status === AppStatuses.active)
+          .map((app) => appToTableRow(app))}
+      </tbody>
+    </Table>
+  )
+}
 
 export default ActiveApps
