@@ -1,29 +1,42 @@
 import * as React from 'react'
-import { FormattedMessage } from 'react-intl'
-import { connect } from 'react-redux'
-import { RootState } from '../ducks'
+import { useTranslation } from 'react-i18next'
+import useLocalStorage from 'use-local-storage'
 import createHistoryItems, { IHistoryItem } from './HistoryItems'
 
-interface IBlog {
-  histories: IHistoryItem[]
+interface IBlogFeed {
+  title: string
+  link: string
+  published: string
+  [attr: string]: string
 }
 
-const Blog: React.SFC<IBlog> = ({ histories }) => (
-  <article className="history">
-    <h4>
-      <FormattedMessage id="navigation.blog" />
-    </h4>
-    <hr />
-    <ul className="list-unstyled">{createHistoryItems(histories)}</ul>
-  </article>
-)
+export default () => {
+  const [histories, setHistories] = useLocalStorage<
+    ReadonlyArray<IHistoryItem>
+  >('blog', [])
+  const { t } = useTranslation()
 
-const mapStateToProps = (state: RootState) => ({
-  histories: state.blog.map((feed) => ({
-    date: new Date(feed.published),
-    title: feed.title,
-    url: feed.link,
-  })),
-})
+  React.useEffect(() => {
+    fetch(
+      'https://wudix076af.execute-api.ap-northeast-1.amazonaws.com/Prod/feed'
+    )
+      .then((res) => res.json())
+      .then((body) =>
+        setHistories(
+          body.data.map((feed: IBlogFeed) => ({
+            date: new Date(feed.published),
+            title: feed.title,
+            url: feed.link,
+          }))
+        )
+      )
+  }, [])
 
-export default connect(mapStateToProps)(Blog)
+  return (
+    <article className="history">
+      <h4>{t('navigation.blog')}</h4>
+      <hr />
+      <ul className="list-unstyled">{createHistoryItems(histories)}</ul>
+    </article>
+  )
+}
