@@ -1,23 +1,14 @@
 import * as React from 'react'
-import { FormattedDate } from 'react-intl'
+import { FormatDate } from './DateTime'
 
 export interface IHistoryItem {
   date: Date
   title: string
   url?: string
 }
-export const HistoryItem: React.SFC<IHistoryItem> = ({ date, title, url }) => {
-  const line = [
-    <FormattedDate
-      key="dummy"
-      value={date}
-      year="numeric"
-      month="2-digit"
-      day="2-digit"
-    />,
-    ' ',
-    title,
-  ]
+
+export const HistoryItem: React.FC<IHistoryItem> = ({ date, title, url }) => {
+  const line = [<FormatDate key={`${date}_${title}`} date={date} />, ' ', title]
   return url ? (
     <li className="history-item">
       <a href={url} target="_blank">
@@ -29,7 +20,7 @@ export const HistoryItem: React.SFC<IHistoryItem> = ({ date, title, url }) => {
   )
 }
 
-export default (histories: IHistoryItem[]) =>
+export default (histories: ReadonlyArray<IHistoryItem>) =>
   histories.map((history) => (
     <HistoryItem key={JSON.stringify(history)} {...history} />
   ))
