@@ -1,12 +1,12 @@
-import * as React from 'react'
 import { withPrefix } from 'gatsby-link'
+import * as React from 'react'
 
 interface ILogoSize {
   width?: string
   height?: string
 }
 
-const Logo: React.SFC<ILogoSize> = (props) => (
+const Logo: React.FC<ILogoSize> = (props) => (
   <img src={withPrefix(`img/logo-white.gif`)} alt="Logo" {...props} />
 )
 
